This is a major upgrade to the services codebase. All classes that revolve around a database record now inherit from a DB_Record class, which contains common record-level DB functions that used to be duplicated in all subsequent classes. Not only does this clean up the code considerably, but it reduces the possibility for bugs in 95% of all DB operations within services.
Additionally, all services now check for account updates every 5 seconds. Any
recent updates are then loaded into memory. This fixes several issues stemming
from a change to an account being made on one service and another service not
being aware of the change...ever.

The operator service now has the capability to gline any Tor hosts. It does
this by comparing each connection to a list of Tor node IP addresses that gets
loaded into memory at startup. A list like this can easily be found by
searching Google. It only needs a text file with one IP/host per line.

A new statistics service has been added, and is still in alpha stages. This
should not be used in production anywhere as it just is not ready yet. The
stats service will keep a real-time database snapshot of channels, users,
servers, and summary data. Its purpose is to provide real-time data to
web-based applications and any other application that, for security or
logistical reasons, should not have access to the services database.

Several services have bug fixes, and also have new commands and/or extensions
to existing commands. The details for these can be found in the ChangeLog, but
here is a quick summary:
        Channel: CHANINFO (new), SET NOPURGE (extension),
                ACCTRELOAD (new), INVITE (event)
        Nickname: SET (admins can now modify other users' settings)

Services are now at the following versions:
        Services Core: 1.1.0
        Channel:       1.5.0
        Nickname:      1.3.6
        Operator:      1.1.0
        Statistics:    1.0.0